# faster speed for decompressSequencesLong