@senhuang42 2021-05-05
What are the scenarios in which this prefetching might be left unused/wasted?