@Cyan4973 2021-05-05
> What are the scenarios in which this prefetching might be left unused/wasted?

The decoder only prefetches memory regions
that are effectively going to be copied later for a match operations.

That being said, in some cases, when a match's offset is really small, 
the source region may not yet be filled (at the time prefetching is issued).

That should not matter much because it means this memory region is very fresh, 
hence likely already in L1, and would have been in L1 anyway, even without prefetching.

So we could say that, in this case, the unconditional prefetching was "useless".
However, branching on the speculative presence of a memory region within L1
is way more expensive than merely prefetching unconditionally inside the hot loop.
So it's preferable to always prefetch, even when a memory region is likely already present in L1.